dashboard/backend/requirements.txt
python-multipart==0.0.6
fastapi==0.115.0
uvicorn[standard]==0.31.0
gunicorn==23.0.0
python-dotenv==1.0.1
supabase==2.9.0
pydantic==2.9.2
google-generativeai==0.8.3
google-genai==0.2.2
Pillow==11.0.0
elevenlabs==2.16.0
requests==2.32.3
anthropic==0.39.0
pyjwt==2.8.0
twilio==9.3.7
ics==0.7.2
python-dateutil==2.9.0
numpy==1.26.4
sentence-transformers==3.3.1
onnxruntime==1.20.1
pydub==0.25.1
